refactor: absorb ClientConnection + resolveConnectionMode into existing files
- Merge /connect, /offline, /reconnect commands into ClientNetwork.init()
- Inline resolveConnectionMode logic directly in client/index.js
- Wire navigateToServer helper into Sidebar.handleConnect
- Delete ClientConnection.js and resolveConnectionMode.js

Net result: 3 new files → 1 (OfflineNetwork.js remains).
