+++ src/client/index.js
@@ -3,22 +3,34 @@ import 'ses'
 import '../core/lockdown'
 import { createRoot } from 'react-dom/client'
 
-import { resolveConnectionMode } from '../core/resolveConnectionMode'
+import { sanitizeWsUrl } from '../core/utils'
 import { OfflineNetwork } from '../core/systems/OfflineNetwork'
 import { Client } from './world-client'
 
 function App() {
   const searchParams = new URLSearchParams(location.search)
   const allowOverride =
-    location.hostname === 'localhost' ||
-    env.PUBLIC_ALLOW_WS_OVERRIDE === 'true'
-  const conn = resolveConnectionMode({
-    wsUrl: env.PUBLIC_WS_URL,
-    searchParams,
-    allowOverride,
-  })
-  const NetworkSystem = conn.mode !== 'direct' ? OfflineNetwork : undefined
-  return <Client wsUrl={conn.wsUrl} networkSystem={NetworkSystem} />
+    location.hostname === 'localhost' || env.PUBLIC_ALLOW_WS_OVERRIDE === 'true'
+
+  let wsUrl = env.PUBLIC_WS_URL
+  let mode = searchParams.get('mode')
+
+  if (mode !== 'offline' && mode !== 'solo') {
+    if (allowOverride) {
+      const connectUrl = searchParams.get('connect')
+      if (connectUrl) {
+        const clean = sanitizeWsUrl(connectUrl)
+        if (clean) {
+          wsUrl = clean
+          mode = 'direct'
+        }
+      }
+    }
+    if (!mode) mode = wsUrl ? 'direct' : 'offline'
+  }
+
+  const NetworkSystem = mode !== 'direct' ? OfflineNetwork : undefined
+  return <Client wsUrl={wsUrl} networkSystem={NetworkSystem} />
 }
 
 const root = createRoot(document.getElementById('root'))
