+++ src/core/createClientWorld.js
@@ -23,7 +23,6 @@ import { Snaps } from './systems/Snaps'
 import { Wind } from './systems/Wind'
 import { XR } from './systems/XR'
 import { ClientAI } from './systems/ClientAI'
-import { ClientConnection } from './systems/ClientConnection'
 
 export function createClientWorld({ NetworkSystem } = {}) {
   const world = new World()
@@ -50,6 +49,5 @@ export function createClientWorld({ NetworkSystem } = {}) {
   world.register('wind', Wind)
   world.register('xr', XR)
   world.register('ai', ClientAI)
-  world.register('connection', ClientConnection)
   return world
 }
