+++ src/core/utils-client.js
@@ -6,6 +6,21 @@
  * carefully does this the same way as the server function.
  *
  */
+/**
+ * Navigate to a server URL by reloading with ?connect=<wsUrl>.
+ * Pass no argument (or undefined) to clear the connect param and use env default.
+ */
+export function navigateToServer(wsUrl) {
+  const url = new URL(location.href)
+  url.searchParams.delete('mode')
+  if (wsUrl) {
+    url.searchParams.set('connect', wsUrl)
+  } else {
+    url.searchParams.delete('connect')
+  }
+  location.href = url.toString()
+}
+
 export async function hashFile(file) {
   const buf = await file.arrayBuffer()
   const hashBuf = await crypto.subtle.digest('SHA-256', buf)
